Fix tout normalement good

+++ ft_calloc.c
@@ -18,7 +18,7 @@ void	*ft_calloc(size_t nmemb, size_t size)
 
 	if ((long long)size == 0 || (long long)nmemb == 0)
 		return (malloc(1));
-	if ((long)size < 0 || (long long)nmemb < 0)
+	if ((long long)size < 0 || (long long)nmemb < 0)
 		return (NULL);
 	tab = malloc(nmemb * size);
 	if (!tab)
